requirements.txt
Django==5.0.1
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
Pillow==10.2.0
python-dotenv==1.0.0 